Yayyy  return empty when empty
The bug was that although that the current index was less than markdown.length, so it did go through looking for the characters [,],(,) , but since it didn't find it would add nothing so the array was as if it din't exist. When it did it was simply empty. (Symptom).  It wouldn't have been caught if it wasn't for the failure-induced input but it allowed us to go back and to the .java file and consider  to code the empty array before trying to add an element that was not present. Because the condition was true it kept looping and looping but since no elements were added, there was no way to escape the loop and thus never returned the array "toReturn". So by writing an if statement in the middle it allowed efficient fdebugging